+++ loadformats.ps1
@@ -1,0 +1,6 @@
+# load the format files
+
+Get-ChildItem $PSScriptroot\*format.ps1xml |
+Foreach-Object {
+    Update-FormatData -AppendPath $_.FullName
+}