+++ myBackupPending.ps1
@@ -9,9 +9,6 @@ Param(
     [ValidateNotNullOrEmpty()]
     [string]$Backup = "*",
 
-    [Parameter(HelpMessage = "Show raw objects, not a formatted report.")]
-    [switch]$Raw,
-
     [Parameter(HelpMessage = "Specify the location for the backup-log.csv files.")]
     [ValidateNotNullOrEmpty()]
     [ValidateScript({Test-Path $_})]
@@ -29,40 +26,17 @@ $f = Get-ChildItem -path $csv | ForEach-Object {
 
 Write-Verbose "Found $($in.count) files to import"
 Write-Verbose "Getting unique file names from $($f.count) files"
-$files = ($f.name | Select-Object -Unique).Foreach( {$n = $_; $f.where( { $_.name -eq $n }) |
+$files = ($f.name | Select-Object -Unique).Foreach({
+    $n = $_;
+    $f.where({ $_.name -eq $n }) |
 Sort-Object -Property {$_.date -as [datetime] } | Select-Object -last 1 })
 
 Write-Verbose "Found $($files.count) unique files"
 
-if ($raw) {
-    Write-Verbose "Displaying raw file data"
-    $files
-}
-else {
-    Write-Verbose "Grouping files"
-    $grouped = $files | Group-Object Log
-
-    Write-Verbose "Display formatted results"
-    $files |
-    Sort-Object -Property Log, Directory, Name |
-    Format-Table -GroupBy log -Property Date, Name, Size, Directory
-
-    $summary = foreach ($item in $grouped) {
-        [PSCustomObject]@{
-            Backup = (Get-Item $item.Name).basename.split("-")[0]
-            Files  = $item.Count
-            Size   = ($item.group | Measure-Object -Property size -sum).sum
-        }
-    } #foreach item
-
-    $total = [PSCustomObject]@{
-        TotalFiles  = ($grouped | Measure-Object -property count -sum).sum
-        TotalSizeMB = [math]::round(($summary.size | Measure-Object -sum).sum / 1MB, 4)
-    }
-    Write-Host "Incremental Pending Backup Summary" -ForegroundColor cyan
-    ($summary | Format-Table | Out-String).TrimEnd() | Write-Host -ForegroundColor cyan
-
-    $total | Format-Table  | Out-String | Write-Host -ForegroundColor cyan
-}
+$files | ForEach-Object {
+    #insert a new typename
+    $_.psobject.typenames.insert(0,'pendingFile')
+    $_
+} | Sort-Object -Property Log, Directory, Name
 
 Write-Verbose "Pending report finished"